requirements.txt
python-dotenv>=1.0.0

# Gmail API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# AI extraction (Phase B)
anthropic>=0.39.0

# Brochure parsing (Phase B)
pdfplumber>=0.11.0
PyMuPDF>=1.24.0

# Excel read/write
openpyxl>=3.1.0
